move `config.py` to root (imports)
